Cleanup
1. Add script block to formlayout.
2. Split util js file
