+++ funnel/assets/js/submissions.js
@@ -1,4 +1,5 @@
-import { Utils, TableSearch } from './util';
+import Form from './utils/formhelper';
+import TableSearch from './utils/tablesearch';
 
 $(() => {
   window.Hasgeek.submissionsInit = function submissionsInit({
@@ -55,7 +56,7 @@ $(() => {
             },
             dataType: 'json',
             error(errorResponse) {
-              Utils.handleAjaxError(errorResponse);
+              Form.handleAjaxError(errorResponse);
               $('.proposal-list-table tbody').sortable('cancel');
             },
           });
