+++ funnel/assets/js/utils/gettext.js
