+++ funnel/assets/js/utils/lazyloadimage.js
@@ -1,0 +1,26 @@
+const LazyloadImg = {
+  init(imgClassName) {
+    this.imgItems = [...document.querySelectorAll(`.${imgClassName}`)];
+    this.imgItems.forEach((img) => {
+      if (img) {
+        let observer = new IntersectionObserver(
+          (entries) => {
+            entries.forEach((entry) => {
+              if (entry.isIntersecting) {
+                entry.target.src = entry.target.dataset.src;
+                observer = observer.disconnect();
+              }
+            });
+          },
+          {
+            rootMargin: '0px',
+            threshold: 0,
+          }
+        );
+        observer.observe(img);
+      }
+    });
+  },
+};
+
+export default LazyloadImg;
