+++ funnel/static/sass/_layout.scss
@@ -60,7 +60,7 @@
 }
 
 .separator {
-  margin: 10px 0;
+  margin: $mui-grid-padding 0;
 }
 
 .flex-wrapper {
