+++ funnel/static/sass/_submission.scss
@@ -244,44 +244,8 @@
   }
 }
 
-.proposal-move-form div {
-  display: inline-block;
-}
-
-.proposal-move-form .hg-form {
-  margin-top: 13px;
-}
-
-.proposal-move-form .hg-form .mui-btn {
-  margin-left: 5px;
-}
-
-.proposal-move-form .select2-container {
-  margin-top: -11px;
-}
-
-.proposal-move-form label {
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 24px;
-  top: -33px;
-  color: $mui-text-dark;
-}
-
 .proposal__admin-panel .mui-select {
   width: 100%;
-}
-
-.proposal__admin-panel .btn-group .mui-btn {
-  margin-bottom: $mui-grid-padding;
-  margin-right: 20px;
-}
-
-@media (min-width: 1200px) {
-  .proposal__admin-panel .btn-group .mui-btn {
-    margin-bottom: 10px;
-    margin-right: 10px;
-  }
 }
 
 .proposal-content {
@@ -323,12 +287,7 @@
   width: 100%;
   position: relative;
   cursor: pointer;
-
-  .material-icons {
-    position: absolute;
-    bottom: 10px;
-    right: 0;
-  }
+  background: $mui-bg-color-primary;
 }
 
 .selected-label--noborder {
@@ -344,7 +303,7 @@
   fieldset {
     border: none;
     position: absolute;
-    top: -100px;
+    top: -50px;
     z-index: -1;
     opacity: 0;
     left: 0px;
