+++ funnel/templates/autoform.html.jinja2
@@ -8,7 +8,7 @@
   {{ renderform(form=form, formid=formid, ref_id=ref_id, submit=submit, message=message, action=action, cancel_url=cancel_url, multipart=multipart, autosave=autosave, draft_revision=draft_revision) }}
 {% endblock %}
 
-{% block footerscripts %}
+{% block innerscripts %}
   {{ widget_ext_scripts(form) }}
   <script type="text/javascript">
     $(function() {
