+++ funnel/templates/labels.html.jinja2
@@ -66,9 +66,4 @@
     <script type="text/javascript" src="{{ ASSET_URL }}"></script>
   {%- endassets -%}
   <script src="{{ built_asset('labels.js') }}" type="text/javascript"></script>
-  <script type="text/javascript">
-    $(function() {
-      window.Hasgeek.labelsInit();
-    });
-  </script>
 {% endblock %}
