+++ funnel/templates/labels_form.html.jinja2
@@ -40,7 +40,7 @@
   {%- if message %}
     <p class="form-message">{{ message }}</p>
   {%- endif %}
-  <form data-parsley-validate="true" id="label-form" method="POST" action="{{ request.url }}" accept-charset="UTF-8" class="mui-form mui-form--margins label-form">
+  <form data-parsley-validate="true" id="{{ ref_id }}" method="POST" action="{{ request.url }}" accept-charset="UTF-8" class="mui-form mui-form--margins label-form">
     {{ form.hidden_tag() }}
     {{ labelformfields(form) }}
     <div id="child-form" class="label-form__subform">
@@ -72,59 +72,19 @@
   </form>
 {% endblock %}
 
-{% block footerscripts %}
+{% block innerscripts %}
   {% assets "js_emojionearea" -%}
     <script type="text/javascript" src="{{ ASSET_URL }}"></script>
   {%- endassets -%}
   {% assets "js_sortable" -%}
     <script type="text/javascript" src="{{ ASSET_URL }}"></script>
   {%- endassets -%}
+  <script src="{{ 'parsley.js'|ext_asset_url }}" type="text/javascript"></script>
+  <script src="{{ built_asset('labels_form.js') }}" type="text/javascript"></script>
   <script type="text/javascript">
     $(function() {
-      function initEmojiPicker() {
-        $('input.field-icon_emoji:not([style*="display: none"])').emojioneArea({
-          pickerPosition: 'bottom',
-          autocomplete: false,
-          standalone: true
-        });
-      }
-      function applySortable() {
-        $(this).sortable({
-          placeholder: $(this).data('drag-placeholder'),
-          cursor: 'move',
-          update: function() {
-              $(this).children().each(function(index) {
-                $(this).children('input[name$="seq"]').val(++index);
-              });
-          }
-        });
-      }
-      $('#add-sublabel-form').click(function(event) {
-        event.preventDefault();
-        var formHtml = {{ labelformfields(emptysubform, subform=true, empty=true)|tojson|safe }};
-        $('#child-form').append(formHtml);
-        activate_widgets();
-        initEmojiPicker();
-        $('.js-required-field').removeClass('mui--hide');
-        $('.js-required-field input').prop('checked', true);
-        $("#child-form").each(applySortable);
-      });
-      $('#child-form').on('click', '.js-remove-sublabel-form', function() {
-        event.preventDefault();
-        $(this).parent().remove();
-      });
-      initEmojiPicker();
-      $("#child-form").each(applySortable);
-
-      $('#label-form').on('submit', function(event) {
-        var optionCount = $("#child-form").find('.ui-draggable-box').length;
-        if(optionCount === 1) {
-          event.preventDefault();
-          window.toastr.error('Minimum 2 or more options are needed');
-          return false;
-        }
-      });
+      var formHtml = {{ labelformfields(emptysubform, subform=true, empty=true)|tojson|safe }};
+      window.Hasgeek.LabelsFormInit(formHtml);
     });
   </script>
-  <script src="{{ 'parsley.js'|ext_asset_url }}" type="text/javascript"></script>
 {% endblock %}
