+++ funnel/templates/project_cfp.html.jinja2
@@ -13,7 +13,7 @@
   {%- if message %}
     <p class="form-message">{{ message }}</p>
   {%- endif %}
-  <form data-parsley-validate="true" id="cfp-form" method="POST" action="{{ request.url }}" accept-charset="UTF-8" class="mui-form mui-form--margins hg-form">
+  <form data-parsley-validate="true" id="{{ ref_id }}" method="POST" action="{{ request.url }}" accept-charset="UTF-8" class="mui-form mui-form--margins hg-form">
     {{ form.hidden_tag() }}
     {{ renderfield(form.instructions, autofocus=true) }}
     {%- for error in form.instructions.errors %}
@@ -44,27 +44,7 @@
   </form>
 {% endblock %}
 
-{% block footerscripts %}
-  <script type="text/javascript">
-    $(function() {
-      function addZero(i) {
-        // input="time" widget requires zero-filled numbers
-        return (i < 10) ? ("0" + i) : i;
-      }
-      // Set current date & time in cfp_start_at input field and show cfp_end_at input field.
-      $('button[name="open-now"]').click(function(event) {
-        event.preventDefault();
-        var now = new Date(new Date().getTime() + new Date().getTimezoneOffset() * -60 * 1000);
-        $('#cfp_start_at').val(now.toISOString().slice(0, 16));
-        $('.js-cfp-start-at, .js-cfp-end-at').removeClass('mui--hide');
-      });
-
-      // Show cfp_start_at and cfp_end_at input field.
-      $('button[name="later"]').click(function(event) {
-        event.preventDefault();
-        $('.js-cfp-start-at, .js-cfp-end-at').removeClass('mui--hide');
-      });
-    });
-  </script>
+{% block innerscripts %}
   <script src="{{ 'parsley.js'|ext_asset_url }}" type="text/javascript"></script>
+  <script src="{{ built_asset('cfp_form.js') }}" type="text/javascript"></script>
 {% endblock %}
