+++ funnel/templates/session_form.html.jinja2
@@ -6,13 +6,13 @@
 {%- endif %}
 
 {% block content %}
-  <form id="{{ formid }}" method="POST" accept-charset="UTF-8" class="mui-form">
+  <form id="{{ ref_id }}" method="POST" accept-charset="UTF-8" class="mui-form" action="{{ request.url }}">
     {{ renderform_inner(form, formid or None) }}
     {%- if not request_is_xhr() %}
       {{ rendersubmit([("Save session", "save session", '')]) }}
     {%- endif %}
   </form>
   {%- if request_is_xhr() %}
-    {{ ajaxform(ref_id=formid, request=request, force=false) }}
+    {{ ajaxform(ref_id=ref_id, request=request, force=false) }}
   {%- endif %}
 {% endblock %}
