+++ funnel/templates/submission_admin_panel.html.jinja2
@@ -5,14 +5,8 @@
   <a class="modal__close mui--text-dark" aria-label="{% trans %}Close{% endtrans %}" rel="modal:close" data-cy="close-admin-panel" href="javascript:void(0);" role="button" tabindex="0">{{ faicon(icon='times', baseline=false, icon_size='title') }}</a>
   <h3 class="mui--text-title mui--text-bold mui--text-uppercase" data-cy="admin-panel">{% trans %}Admin panel{% endtrans %}</h3>
 </div>
-<div class="modal__body">
+<div class="modal__body proposal__admin-panel">
   {%- if proposal.current_roles.project_editor %}
-    {%- if proposal_move_form %}
-      {{ renderform(proposal_move_form, 'proposal-move', action=proposal.url_for('moveto'), submit=gettext("Move"), style='horiz') }}
-    {%- endif %}
-    {%- if proposal_transfer_form %}
-      {{ renderform(proposal_transfer_form, 'proposal-transfer', action=proposal.url_for('transfer_to'), submit=gettext("Transfer")) }}
-    {%- endif %}
     <p class="mui--text-subhead">{% trans %}Change status of proposal{% endtrans %}</p>
     <form action="{{ proposal.url_for('transition') }}" method="post" class="form-inline">
       {{ transition_form.hidden_tag() }}
@@ -23,11 +17,9 @@
       </p>
     </form>
   {% endif %}
-  {% if project.view_for('labels').is_available() %}
+  {% if project.view_for('labels').is_available() and proposal_label_admin_form %}
     <hr>
     <p class="mui--text-subhead"><a href="{{ project.url_for('labels') }}">{{ faicon(icon='tag', css_class='fa-icon--right-margin mui--align-middle') }}{% trans %}Manage labels{% endtrans %}</a></p>
-  {% endif %}
-  {% if proposal_label_admin_form %}
     <form action="{{ proposal.url_for('edit_labels') }}" method="post" class="add-label-form">
       {{ proposal_label_admin_form.hidden_tag() }}
       <p class="mui--text-title">Labels</p>
@@ -43,6 +35,12 @@
       {{ rendersubmit([("add-label", "Save", '')]) }}
     </form>
   {% endif %}
+  {%- if proposal_move_form %}
+    {{ renderform(proposal_move_form, 'proposal-move', action=proposal.url_for('moveto'), submit=gettext("Move")) }}
+  {%- endif %}
+  {%- if proposal_transfer_form %}
+    {{ renderform(proposal_transfer_form, 'proposal-transfer', action=proposal.url_for('transfer_to'), submit=gettext("Transfer")) }}
+  {%- endif %}
 </div>
 
 {{ widgetscripts(proposal_transfer_form, script=true) }}
